packet-capture/requirements.txt
scapy==2.5.0
requests==2.31.0
python-dotenv==1.0.0
psutil==5.9.0
colorama==0.4.6
